refactor to make index less busy. add local env vars
